Cleanup components to have index and clear hierarchy (#154)
