+++ config/webpack.client.config.js
@@ -137,6 +137,7 @@ module.exports = {
   ],
   resolve: {
     extensions: ['*', '.js', '.jsx'],
+    alias: { 'react-dom': '@hot-loader/react-dom' },
   },
 };
