+++ src/components/action-bar/action-bar.styles.js
@@ -3,8 +3,11 @@ const ActionBarStyles = theme => ({
     backgroundColor: theme.palette.common.white,
     color: theme.palette.common.black,
   },
-  closeButton: {
+  navButton: {
     marginRight: theme.spacing(2),
+  },
+  actionButtonGroup: {
+    marginRight: theme.spacing(4),
   },
   title: {
     flexGrow: 1,
