+++ src/components/action-bar/components/profile-menu/profile-menu.styles.js
