+++ src/components/action-bar/components/profile-popover/profile-popover.jsx
