+++ src/components/action-bar/index.js
@@ -1,3 +1,8 @@
-import ActionBar from './action-bar.component';
+import ActionBar from './action-bar';
+import ElevationScroll from './components/elevation-scroll/elevation-scroll';
+import ProfileMenu from './components/profile-menu/profile-menu';
+import ProfilePopover from './components/profile-popover/profile-popover';
 
-export default ActionBar;
+export {
+  ActionBar, ElevationScroll, ProfileMenu, ProfilePopover,
+};
