+++ src/components/alert-dialog/alert-dialog.jsx
@@ -43,6 +43,7 @@ const AlertDialog = (props) => {
         <Button
           onClick={handleConfirm}
           color="primary"
+          variant="contained"
           autoFocus
           disabled={isDisabled}
         >
