+++ src/components/alert-dialog/index.js
@@ -1,0 +1,1 @@
+export { default } from './alert-dialog';
