--- src/components/entity-form-action-buttons/entity-form-action-buttons.styles.js
@@ -1,7 +1,0 @@
-const EntityFormActionButtonsStyles = theme => ({
-  button: {
-    margin: theme.spacing(1),
-  },
-});
-
-export default EntityFormActionButtonsStyles;
