--- src/components/entity-grid/entity-grid.styles.js
@@ -1,15 +1,0 @@
-const EntityGridStyles = theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  grid: {
-    width: '100%',
-  },
-  paper: {
-    padding: theme.spacing(1),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-  },
-});
-
-export default EntityGridStyles;
