+++ src/components/full-drawer/index.js
@@ -1,0 +1,5 @@
+import FullDrawer from './full-drawer';
+import FullDrawerHeader from './components/full-drawer-header/full-drawer-header';
+import FullDrawerListItem from './components/full-drawer-list-item/full-drawer-list-item';
+
+export { FullDrawer, FullDrawerHeader, FullDrawerListItem };
