+++ src/components/google-login-button/index.js
@@ -1,0 +1,1 @@
+export { default } from './google-login-button';
