+++ src/components/group-grid/components/group-item/group-item.jsx
