+++ src/components/group-grid/group-grid.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Grid, Typography } from '@material-ui/core';
-import GroupItem from './components/group-item/group-item.component';
+import GroupItem from './components/group-item/group-item';
 
 const GroupGrid = (props) => {
   const { groups, isRemovable } = props;
