+++ src/components/group-grid/index.js
@@ -1,0 +1,4 @@
+import GroupGrid from './group-grid';
+import GroupItem from './components/group-item/group-item';
+
+export { GroupGrid, GroupItem };
