--- src/components/group-grid/index.jsx
@@ -1,4 +1,0 @@
-import GroupGrid from './group-grid.component';
-import GroupItem from './components/group-item/group-item.component';
-
-export { GroupGrid, GroupItem };
