--- src/components/header/header.styles.js
@@ -1,31 +1,0 @@
-const HeaderStlyes = theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  appBar: {
-    backgroundColor: theme.palette.common.white,
-    color: theme.palette.common.black,
-    transition: theme.transitions.create(['margin', 'width'], {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.leavingScreen,
-    }),
-  },
-  hide: {
-    display: 'none',
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  title: {
-    flexGrow: 1,
-  },
-  iconButton: {
-    padding: 0,
-    backgroundColor: theme.palette.primary,
-  },
-  avatar: {
-    margin: 0,
-  },
-});
-
-export default HeaderStlyes;
