+++ src/components/image-grid/index.js
@@ -1,4 +1,4 @@
-import ImageGrid from './image-grid.component';
-import ImageGridItem from './components/image-grid-item/image-grid-item.component';
+import ImageGrid from './image-grid';
+import ImageGridItem from './components/image-grid-item/image-grid-item';
 
 export { ImageGrid, ImageGridItem };
