+++ src/components/mini-drawer/components/mini-drawer-list-item/mini-drawer-list-item.jsx
