+++ src/components/mini-drawer/components/mini-drawer-list-item/mini-drawer-list-item.styles.js
@@ -1,4 +1,4 @@
-import { miniDrawerWidth } from '../../utils/theme';
+import { miniDrawerWidth } from '../../../../utils/theme';
 
 const MiniDrawerListItemStyles = theme => ({
   link: {
