+++ src/components/mini-drawer/index.js
@@ -1,0 +1,4 @@
+import MiniDrawer from './mini-drawer';
+import MiniDrawerListItem from './components/mini-drawer-list-item/mini-drawer-list-item';
+
+export { MiniDrawer, MiniDrawerListItem };
