+++ src/components/snackbar-content/index.js
@@ -1,0 +1,1 @@
+export { default } from './snackbar-content';
