+++ src/components/snackbar-content/snackbar-content.jsx
