--- src/pages/entity-list/entity-list.styles.js
@@ -1,7 +1,0 @@
-const EntityListStyles = {
-  progressBarContainer: {
-    flexGrow: 1,
-  },
-};
-
-export default EntityListStyles;
