--- src/utils/constants.js
@@ -1,6 +1,0 @@
-const EntityType = Object.freeze({
-  IMAGE: 'image',
-  GROUP: 'group',
-});
-
-export default EntityType;
